// src/enums/message.ts
export enum Messages {
  CREATED = "created successfully",
  UPDATED = "updated successfully",
  DELETED = "deleted successfully",
  FOUND = "found successfully",
  FETCHED = "fetched successfully",
  FAILED = "Something went wrong",
  INVALID = "Invalid data",
  AUTH_SUCCESS = "Authentication successful",
  AUTH_FAILED = "Authentication failed",
}